# Improve perf by using --disable-gems