--- a/src/parse.js
+++ b/src/parse.js
@@ -2,7 +2,7 @@ const { spawnSync } = require("child_process");
 const path = require("path");
 
 module.exports = (text, _parsers, _opts) => {
-  const child = spawnSync("ruby", [path.join(__dirname, "./ripper.rb")], {
+  const child = spawnSync("ruby", ["--disable-gems", path.join(__dirname, "./ripper.rb")], {
     input: text,
     maxBuffer: 10 * 1024 * 1024 // 10MB
   });