--- a/test/js/setupTests.js
+++ b/test/js/setupTests.js
@@ -4,7 +4,7 @@ const prettier = require("prettier");
 const readline = require("readline");
 
 // Set RUBY_VERSION so certain tests only run for certain versions
-process.env.RUBY_VERSION = spawnSync("ruby", ["-e", "puts RUBY_VERSION"])
+process.env.RUBY_VERSION = spawnSync("ruby", ["--disable-gems", "-e", "puts RUBY_VERSION"])
   .stdout.toString()
   .trim();