Loading rubygems can be a slow operation. Worse yet, the load time
performance degrades as more gems are installed on the system. Luckily,
Ruby provides a flag to fix this: `--disable-gems`.

Using this flag significantly improves performance on my machine, here
are some results.

### Without --disable-gems

| file                                             | lint time |
|--------------------------------------------------|-----------|
| apps/admin/application.rb                        | 215ms     |
| apps/admin/config/routes.rb                      | 165ms     |
| apps/admin/controllers/about/show.rb             | 149ms     |
| apps/admin/controllers/accounts/index.rb         | 147ms     |
| apps/admin/controllers/authentication_helpers.rb | 151ms     |
| apps/admin/controllers/deployments/index.rb      | 148ms     |
| apps/admin/controllers/docs/index.rb             | 147ms     |
| apps/admin/controllers/docs/show.rb              | 144ms     |

### With --disable-gems

| file                                             | lint time |
|--------------------------------------------------|-----------|
| apps/admin/application.rb                        | 140ms     |
| apps/admin/config/routes.rb                      | 90ms      |
| apps/admin/controllers/about/show.rb             | 94ms      |
| apps/admin/controllers/accounts/index.rb         | 89ms      |
| apps/admin/controllers/authentication_helpers.rb | 92ms      |
| apps/admin/controllers/deployments/index.rb      | 87ms      |
| apps/admin/controllers/docs/index.rb             | 84ms      |
| apps/admin/controllers/docs/show.rb              | 88ms      |