@kddeisz 2020-02-26
Let's goooo! This is such a great PR nice work!

Let's get this over the finish line. Looks like `prettier` needs to be run on `src/parse.js` and `test/js/setupTests.js`, which is why it's failing CI. Other than that, looks great!